# write code for fifth test, change variable type to int

+++ src/test/java/interstingNumber/CarMileageTest.java
@@ -7,41 +7,41 @@ public class CarMileageTest {
 
   @Test
   public void should_return_0_when_check_interesting_number_given_number_0() {
-    long inputNumber = 0;
-    long expectedResult = 0;
-    long actualResult = CarMileage.isInteresting(inputNumber);
+    int inputNumber = 0;
+    int expectedResult = 0;
+    int actualResult = CarMileage.isInteresting(inputNumber);
     Assert.assertEquals(expectedResult, actualResult);
   }
 
   @Test
   public void should_return_0_when_check_interesting_number_given_number_1() {
-    long inputNumber = 1;
-    long expectedResult = 0;
-    long actualResult = CarMileage.isInteresting(inputNumber);
+    int inputNumber = 1;
+    int expectedResult = 0;
+    int actualResult = CarMileage.isInteresting(inputNumber);
     Assert.assertEquals(expectedResult, actualResult);
   }
 
   @Test
   public void should_return_0_when_check_interesting_number_given_number_digit_followed_by_all_zeros_100() {
-    long inputNumber = 100;
-    long expectedResult = 1;
-    long actualResult = CarMileage.isInteresting(inputNumber);
+    int inputNumber = 100;
+    int expectedResult = 1;
+    int actualResult = CarMileage.isInteresting(inputNumber);
     Assert.assertEquals(expectedResult, actualResult);
   }
 
   @Test
   public void should_return_0_when_check_interesting_number_given_number_digits_are_sequential_and_incement_123() {
-    long inputNumber = 123;
-    long expectedResult = 1;
-    long actualResult = CarMileage.isInteresting(inputNumber);
+    int inputNumber = 123;
+    int expectedResult = 1;
+    int actualResult = CarMileage.isInteresting(inputNumber);
     Assert.assertEquals(expectedResult, actualResult);
   }
 
   @Test
   public void should_return_0_when_check_interesting_number_given_number_digits_are_sequential_and_incement_890() {
-    long inputNumber = 890;
-    long expectedResult = 1;
-    long actualResult = CarMileage.isInteresting(inputNumber);
+    int inputNumber = 890;
+    int expectedResult = 1;
+    int actualResult = CarMileage.isInteresting(inputNumber);
     Assert.assertEquals(expectedResult, actualResult);
   }
 }
